Update CDB generation method for better integration with IDEs
Previously the working directory of the compilation command was set to
the execution root. However, this method effectively maps a
compilation method to a file that is outside the project root (the bazel
execution root is usually in a cache directory), which makes the
cross-file renaming and other IDE features not work well.

This commit changes the following aspects in generation of compilation
database:

1. Use `cquery` instead of `query` when fetching the build targets of
   cc_* type. `cquery` uses the correct configuration when making the
   queries and I have found that it works better than the `qeury` command.
2. Set the working directory of the compilation command to the workspace
   directory (`bazel info workspace`, i.e., the project root). To make
   the compiler able to find headers in the `external` folder, I have
   modified the compilation database to the use absolute path of the
   `external` folder.
3. Do not include compilation commands of dependency files in the
   generated compile_commands.json file.